Modified the Modules files moved databaseModule, viewModelModule and RepositoryModule
